task/solutions
Move the solutions to hints folder instead of overwritten the exercise file
